Refactor the filename determination
