+++ abstract_ranker/utils.py
@@ -1,20 +1,6 @@
-from pathlib import Path
-from typing import Any, Dict, Generator, TypeVar
+from typing import Generator, TypeVar
 
 from rich.progress import Progress
-
-
-def generate_ranking_csv_filename(event: Dict[str, Any]) -> Path:
-    """Build a CSV filename for the summary of the abstracts and ranking.
-
-        Format: "<year>-<monday>-<day> <title>.csv"
-    Args:
-        event (Dict[str, Any]): The indico data from the event
-
-    Returns:
-        str: valid filename for the CSV file
-    """
-    return Path(f"{event['startDate']['date']} - {event['title']}.csv")
 
 
 def as_a_number(interest: str) -> int:
